+++ sdk/resourcemanager/hybridkubernetes/armhybridkubernetes/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.1.0-beta.1 (2022-05-31)
+### Features Added
+
+- New const `PrivateLinkStateDisabled`
+- New const `PrivateLinkStateEnabled`
+- New function `PossiblePrivateLinkStateValues() []PrivateLinkState`
+- New field `PrivateLinkScopeResourceID` in struct `ConnectedClusterProperties`
+- New field `PrivateLinkState` in struct `ConnectedClusterProperties`
+
+
 ## 1.0.0 (2022-05-18)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/hybridkubernetes/armhybridkubernetes` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 1.0.0, which contains breaking changes.